requirements.txt
streamlit
python-dotenv
pandas
openpyxl
xlrd>=2.0.1
